Remove PC heuristic type from FAS-based algorithms.
Removed the `PcHeuristicType` functionality across various classes, including associated fields, methods, and references. This simplifies the codebase and eliminates unused heuristic-related functionality within the FAS and related algorithms.
