+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/Fas.java
@@ -1,41 +1,16 @@
-/// ////////////////////////////////////////////////////////////////////////////
-// For information as to what this class does, see the Javadoc, below.       //
-// Copyright (C) 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005, 2006,       //
-// 2007, 2008, 2009, 2010, 2014 by Peter Spirtes, Richard Scheines, Joseph   //
-// Ramsey, and Clark Glymour.                                                //
-//                                                                           //
-// This program is free software; you can redistribute it and/or modify      //
-// it under the terms of the GNU General Public License as published by      //
-// the Free Software Foundation; either version 2 of the License, or         //
-// (at your option) any later version.                                       //
-//                                                                           //
-// This program is distributed in the hope that it will be useful,           //
-// but WITHOUT ANY WARRANTY; without even the implied warranty of            //
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the             //
-// GNU General Public License for more details.                              //
-//                                                                           //
-// You should have received a copy of the GNU General Public License         //
-// along with this program; if not, write to the Free Software               //
-// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
-/// ////////////////////////////////////////////////////////////////////////////
-
 package edu.cmu.tetrad.search;
 
 import edu.cmu.tetrad.data.Knowledge;
-import edu.cmu.tetrad.graph.*;
+import edu.cmu.tetrad.graph.EdgeListGraph;
+import edu.cmu.tetrad.graph.Graph;
+import edu.cmu.tetrad.graph.GraphUtils;
+import edu.cmu.tetrad.graph.Node;
 import edu.cmu.tetrad.search.test.IndependenceResult;
-import edu.cmu.tetrad.search.utils.PcCommon;
 import edu.cmu.tetrad.search.utils.SepsetMap;
 import edu.cmu.tetrad.util.ChoiceGenerator;
-import edu.cmu.tetrad.util.TetradLogger;
 
-import javax.sound.midi.SysexMessage;
 import java.io.PrintStream;
 import java.util.*;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.Future;
 
 /**
  * Implements the Fast Adjacency Search (FAS), which is the adjacency search of the PC algorithm (see). This is a useful
@@ -66,163 +41,53 @@ import java.util.concurrent.Future;
  * @see Fci
  * @see Knowledge
  */
-public class Fas implements IFas {
+public class Fas {
     /**
-     * The test to be used for conditional independence tests.
+     * Represents the core independence test utilized in the FAS algorithm to determine conditional independence
+     * relationships between variables. This test serves as the foundation for building and refining graphs during the
+     * search process.
      */
     private final IndependenceTest test;
     /**
-     * The logger.
+     * A separation set map used to store the results of conditional independence tests performed during the graph
+     * search process. It holds information about the sets of variables that separate pairs of nodes in the graph.
      */
-    private final TetradLogger logger = TetradLogger.getInstance();
+    private final SepsetMap sepset = new SepsetMap();
     /**
-     * The knowledge.
+     * A Knowledge object used to define constraints, such as forbidden or required edges, and other background
+     * information necessary for the conditional independence search process.
+     * <p>
+     * This variable is integral to guiding various search methods and algorithms in the parent class, ensuring that the
+     * results adhere to predetermined conditions or limitations imposed by the user or the context of the analysis.
      */
     private Knowledge knowledge = new Knowledge();
     /**
-     * The sepsets that were discovered in the search.
+     * Specifies the maximum depth to be considered during the conditional independence search process. The depth
+     * determines the maximum number of conditioning variables used in independence tests, where a value of -1 indicates
+     * that no limit is imposed.
      */
-    private SepsetMap sepset = new SepsetMap();
+    private int depth = -1;
     /**
-     * The heuristic to use.
-     */
-    private PcCommon.PcHeuristicType heuristic = PcCommon.PcHeuristicType.NONE;
-    /**
-     * The depth of the search.
-     */
-    private int depth = 1000;
-    /**
-     * Whether the stable adjacency search should be used.
+     * A flag indicating whether the search process will use a stable strategy. When set to {@code true}, the search
+     * follows a stable approach. This flag may influence the behavior or results of certain algorithms in the
+     * conditional independence graph search process.
      */
     private boolean stable = true;
     /**
-     * The elapsed time of the search.
-     */
-    private long elapsedTime = 0L;
-    /**
-     * Whether verbose output should be printed.
-     */
-    private transient PrintStream out = System.out;
-    /**
-     * Whether verbose output should be printed.
+     * A flag indicating whether verbose output is enabled during the search process. If true, additional logging or
+     * debugging information is printed to the configured output stream. If false, only essential information is
+     * displayed.
      */
     private boolean verbose = false;
     /**
-     * Represents the start time of a specific operation or process within the algorithm's execution. This variable is
-     * used to measure elapsed time or to enforce time constraints during the execution.
+     * The output stream used for logging or displaying messages during the search process. By default, it is set to
+     * {@code System.out}. This stream can be replaced or customized by the user to redirect output messages to
+     * alternative destinations.
      */
-    private long startTime;
-    /**
-     * Specifies the timeout value for the search operation or algorithm execution. The timeout value is used to
-     * determine the maximum allowed duration for the execution, preventing any operation from exceeding this limit.
-     */
-    private long timeout;
+    private PrintStream out = System.out;
 
-    /**
-     * Constructor.
-     *
-     * @param test The test to use for oracle conditional independence test results.
-     */
     public Fas(IndependenceTest test) {
         this.test = test;
-    }
-
-    /**
-     * Checks if there is an adjacency between two nodes on a side.
-     *
-     * @param scores      The map of scores for each edge.
-     * @param test        The independence test to use.
-     * @param adjacencies The map of nodes and their adjacent nodes.
-     * @param depth       The maximum depth to search for adjacencies.
-     * @param x           The first node.
-     * @param y           The second node.
-     * @param mainThread  The main thread. Needed to try to interrupt the search if it takes too long.
-     * @return True if there is an adjacency between x and y at the given depth, false otherwise.
-     * @throws InterruptedException if any
-     */
-    private static boolean checkSide(Map<Edge, Double> scores, IndependenceTest test, Map<Node, Set<Node>> adjacencies,
-                                     int depth, Node x, Node y,
-                                     PcCommon.PcHeuristicType heuristic,
-                                     Knowledge knowledge,
-                                     SepsetMap sepsets, Thread mainThread) throws InterruptedException {
-        if (!adjacencies.get(x).contains(y)) return false;
-
-        List<Node> _adjx = new ArrayList<>(adjacencies.get(x));
-        _adjx.remove(y);
-
-        if (heuristic == PcCommon.PcHeuristicType.HEURISTIC_1 || heuristic == PcCommon.PcHeuristicType.HEURISTIC_2) {
-            Collections.sort(_adjx);
-        }
-
-        List<Node> ppx = possibleParents(x, _adjx, knowledge, y);
-
-        Map<Node, Double> scores2 = new HashMap<>();
-
-        for (Node node : ppx) {
-            if (Thread.currentThread().isInterrupted()) {
-                throw new InterruptedException();
-            }
-
-            Double _score = scores.get(Edges.undirectedEdge(node, x));
-            scores2.put(node, _score);
-        }
-
-        if (heuristic == PcCommon.PcHeuristicType.HEURISTIC_3) {
-            ppx.sort(Comparator.comparing(scores2::get));
-            Collections.reverse(ppx);
-        }
-
-        if (ppx.size() >= depth) {
-            ChoiceGenerator cg = new ChoiceGenerator(ppx.size(), depth);
-            int[] choice;
-
-            while ((choice = cg.next()) != null) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                Set<Node> Z = GraphUtils.asSet(choice, ppx);
-
-                // Check for main thread interruption
-                if (mainThread.isInterrupted()) {
-                    Thread.currentThread().interrupt(); // Propagate to current thread
-                    throw new InterruptedException(); // Stop processing
-                }
-
-                // Check for current thread interruption
-                if (Thread.currentThread().isInterrupted()) {
-                    mainThread.interrupt(); // Propagate back to main thread
-                    throw new InterruptedException();
-                }
-
-                boolean independent = test.checkIndependence(x, y, Z).isIndependent();
-
-                // Check for main thread interruption
-                if (mainThread.isInterrupted()) {
-                    Thread.currentThread().interrupt(); // Propagate to current thread
-                    throw new InterruptedException(); // Stop processing
-                }
-
-                // Check for current thread interruption
-                if (Thread.currentThread().isInterrupted()) {
-                    mainThread.interrupt(); // Propagate back to main thread
-                    throw new InterruptedException();
-                }
-
-                boolean noEdgeRequired = knowledge.noEdgeRequired(x.getName(), y.getName());
-
-                if (independent && noEdgeRequired) {
-                    adjacencies.get(x).remove(y);
-                    adjacencies.get(y).remove(x);
-
-                    sepsets.set(x, y, Z);
-
-                    return true;
-                }
-            }
-        }
-
-        return false;
     }
 
     /**
@@ -271,344 +136,182 @@ public class Fas implements IFas {
     }
 
     /**
-     * Performs a search to discover all adjacencies in the graph. The procedure is to remove edges in the graph which
-     * connect pairs of variables that are independent, conditional on some other set of variables in the graph (the
-     * "sepset"). These edges are removed in tiers. First, edges which are independent conditional on zero other
-     * variables are removed, then edges which are independent conditional on one other variable are removed. Then two,
-     * then three, and so on, until no more edges can be removed from the graph. The edges which remain in the graph
-     * after this procedure are the adjacencies in the data.
+     * Performs a conditional independence graph search using the default set of variables from the initialized
+     * independence test. The method delegates the search process to another method that takes a list of variables as
+     * input.
      *
-     * @return An undirected graph that summarizes the conditional independencies in the data. This graph
-     *         represents the adjacencies that remain after applying conditional independence tests.
-     * @throws InterruptedException if any
+     * @return A graph where edges are updated based on the results of conditional independence tests and other provided
+     * constraints.
+     * @throws InterruptedException If the search process is interrupted.
      */
-    @Override
     public Graph search() throws InterruptedException {
         return search(test.getVariables());
     }
 
     /**
-     * Discovers all adjacencies in data.  The procedure is to remove edges in the graph which connect pairs of
-     * variables which are independent, conditional on some other set of variables in the graph (the "sepset"). These
-     * are removed in tiers.  First, edges which are independent, conditional on zero other variables are removed. Then
-     * edges which are independent conditional on one other variable are removed, then two, then three, and so on, until
-     * no more edges can be removed from the graph.  The edges which remain in the graph after this procedure are the
-     * adjacencies in the data.
+     * Searches for conditional independence relationships in a graph constructed from the given list of nodes. Edges
+     * are removed based on the provided set of constraints and the maximum depth for the search.
      *
-     * @param nodes A list of nodes to search over.
-     * @return An undirected graph that summarizes the conditional independencies in the data.
-     * @throws InterruptedException if any
+     * @param nodes The list of nodes to construct the graph and perform the search on.
+     * @return A graph with edges updated based on the search process and conditional independence tests.
+     * @throws InterruptedException If variables provided are not a subset of the ones in the test or if the search
+     *                              process is interrupted.
      */
     public Graph search(List<Node> nodes) throws InterruptedException {
-        Thread mainThread = Thread.currentThread();
-
-        if (startTime <= 0) {
-            startTime = System.currentTimeMillis();
+        if (!new HashSet<>(test.getVariables()).containsAll(nodes)) {
+            throw new InterruptedException("Variables should be a subset of the ones in the test.");
         }
 
-        nodes = new ArrayList<>(nodes);
-
-        this.logger.addOutputStream(out);
-
-        if (verbose) {
-            this.logger.log("Starting Fast Adjacency Search.");
-        }
-
-        this.test.setVerbose(this.verbose);
-
-        int _depth = this.depth;
-
-        if (_depth == -1) {
-            _depth = 1000;
-        }
-
-        this.sepset = new SepsetMap();
-
-        List<Edge> edges = new ArrayList<>();
-
-        if (this.heuristic == PcCommon.PcHeuristicType.HEURISTIC_1) {
-            Collections.sort(nodes);
-        }
-
-        for (int i = 0; i < nodes.size(); i++) {
-            for (int j = i + 1; j < nodes.size(); j++) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                edges.add(Edges.undirectedEdge(nodes.get(i), nodes.get(j)));
-            }
-        }
-
-        Map<Edge, Double> scores = new HashMap<>();
-
-        class Task implements Callable<Map<Edge, Double>> {
-            private final Edge edge;
-            private final IndependenceTest test;
-
-            public Task(Edge edge, IndependenceTest test) {
-                this.edge = edge;
-                this.test = test;
-            }
-
-            @Override
-            public Map<Edge, Double> call() {
-                IndependenceResult result;
-                try {
-                    result = this.test.checkIndependence(edge.getNode1(), edge.getNode2(), new HashSet<>());
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
-
-                Map<Edge, Double> map = new HashMap<>();
-                map.put(edge, result.getScore());
-
-                return map;
-            }
-        }
-
-        List<Task> tasks = new ArrayList<>();
-
-        for (Edge edge : edges) {
-            tasks.add(new Task(edge, test));
-        }
-
-        if (stable) {
-            int parallelism = Runtime.getRuntime().availableProcessors();
-            ForkJoinPool pool = new ForkJoinPool(parallelism);
-            List<Future<Map<Edge, Double>>> theseResults;
-
-            theseResults = pool.invokeAll(tasks);
-
-            for (Future<Map<Edge, Double>> future : theseResults) {
-                try {
-                    Map<Edge, Double> result = future.get();
-
-                    for (Edge edge : result.keySet()) {
-                        scores.put(edge, result.get(edge));
-                    }
-                } catch (InterruptedException e) {
-                    pool.shutdown();
-                    throw e;
-                } catch (ExecutionException e) {
-                    pool.shutdown();
-                    throw new RuntimeException("Possible singularities or NaNs: " + e.getCause());
-                }
-            }
-
-            pool.shutdown();
-        } else {
-            System.out.println("Stable option false");
-
-            for (Task task : tasks) {
-                long current = System.currentTimeMillis();
-
-                if (getTimeout() > 0 && current > getStartTime() + getTimeout()) {
-                    throw new RuntimeException("Timeout time = " + current + " startTime = " + getStartTime()
-                                               + " timeout = " + getTimeout());
-                }
-
-                Map<Edge, Double> result = task.call();
-
-                for (Edge edge : result.keySet()) {
-                    scores.put(edge, result.get(edge));
-                }
-            }
-        }
-
-        if (this.heuristic == PcCommon.PcHeuristicType.HEURISTIC_2 || this.heuristic == PcCommon.PcHeuristicType.HEURISTIC_3) {
-            edges.sort(Comparator.comparing(scores::get));
-        }
-
-        Map<Node, Set<Node>> adjacencies = new HashMap<>();
-
-        for (Node node : nodes) {
-            Set<Node> set = new LinkedHashSet<>();
-
-            for (Node _node : nodes) {
-                if (_node == node) continue;
-                set.add(_node);
-            }
-
-            adjacencies.put(node, set);
-        }
-
-        for (Edge edge : new ArrayList<>(edges)) {
-            if ((scores.get(edge) != null && scores.get(edge) < 0) || (this.knowledge.isForbidden(edge.getNode1().getName(), edge.getNode2().getName())
-                    && (this.knowledge.isForbidden(edge.getNode2().getName(), edge.getNode1().getName())))) {
-                edges.remove(edge);
-                adjacencies.get(edge.getNode1()).remove(edge.getNode2());
-                adjacencies.get(edge.getNode2()).remove(edge.getNode1());
-                this.sepset.set(edge.getNode1(), edge.getNode2(), new HashSet<>());
-            }
-        }
-
-        for (int d = 0; d <= _depth; d++) {
-            if (verbose) {
-                System.out.println("Depth: " + d);
-            }
-
-            boolean more;
-
-            if (this.stable) {
-                Map<Node, Set<Node>> adjacenciesCopy = new HashMap<>();
-
-                for (Node node : adjacencies.keySet()) {
-                    adjacenciesCopy.put(node, new LinkedHashSet<>(adjacencies.get(node)));
-                }
-
-                adjacencies = adjacenciesCopy;
-            }
-
-            more = searchAtDepth(scores, edges, this.test, adjacencies, d, mainThread);
-
-            if (!more) {
-                break;
-            }
-        }
-
-        // The search graph. It is assumed going in that all the true adjacencies of x are in this graph for every node
-        // x. It is hoped (i.e., true in the large sample limit) that true adjacencies are never removed.
         Graph graph = new EdgeListGraph(nodes);
+        graph = GraphUtils.completeGraph(graph);
 
         for (int i = 0; i < nodes.size(); i++) {
             for (int j = i + 1; j < nodes.size(); j++) {
                 Node x = nodes.get(i);
                 Node y = nodes.get(j);
 
-                if (adjacencies.get(x).contains(y)) {
-                    graph.addUndirectedEdge(x, y);
+                if (knowledge.isForbidden(x.getName(), y.getName()) && knowledge.isForbidden(y.getName(), x.getName())) {
+                    graph.removeEdge(x, y);
                 }
             }
         }
 
-        if (verbose) {
-            this.logger.log("Finishing Fast Adjacency Search.");
-        }
+        int depth_ = depth >= 0 ? depth : Math.max(depth, test.getVariables().size() - 1);
 
-        this.elapsedTime = System.currentTimeMillis() - startTime;
+        for (int d = 0; d <= depth_; d++) {
+            if (verbose) {
+                System.out.println("Depth: " + d);
+            }
+
+            Graph graph_ = graph;
+
+            if (this.stable) {
+                graph = new EdgeListGraph(graph_);
+
+                if (searchAtDepth(graph, graph_, d, this.stable)) {
+                    graph = graph_;
+                    break;
+                }
+
+                graph = graph_;
+            } else {
+                if (searchAtDepth(graph, graph_, d, stable)) {
+                    break;
+                }
+            }
+        }
 
         return graph;
     }
 
     /**
-     * Sets the maximum depth for the search.
+     * Searches for conditional independence relationships in a graph up to a given depth d. If the graph's nodes meet
+     * the conditions specified, the edges may be removed based on the provided knowledge.
      *
-     * @param depth The maximum depth to set.
-     * @throws IllegalArgumentException if the depth is less than -1.
+     * @param graph  The graph on which the search is performed.
+     * @param graph_ A copy of the graph where changes are applied during the search process.
+     * @param d      The maximum depth to consider for conditional independence tests.
+     * @param stable If true, performs the search using a parallel strategy; otherwise, uses a sequential strategy.
+     * @return True if the maximum free degree of the resulting graph is less than or equal to d, false otherwise.
+     * @throws InterruptedException if the search is interrupted during independence testing or parent determination.
      */
-    public void setDepth(int depth) {
-        if (depth < -1) {
-            throw new IllegalArgumentException("Depth must be -1 (unlimited) or >= 0.");
+    private boolean searchAtDepth(Graph graph, Graph graph_, int d, boolean stable) throws InterruptedException {
+        List<Node> nodes = graph.getNodes();
+
+
+        if (stable) {
+            nodes.parallelStream().forEach(x -> {
+                for (Node y : nodes) {
+                    if (x == y || nodes.indexOf(x) >= nodes.indexOf(y)) continue;
+
+                    if (knowledge.isForbidden(x.getName(), y.getName()) && knowledge.isForbidden(y.getName(), x.getName())) {
+                        synchronized (graph_) {
+                            graph_.removeEdge(x, y);
+                        }
+                    }
+
+                    List<Node> adjX = graph.getAdjacentNodes(x);
+                    adjX.remove(y);
+
+                    List<Node> ppx = null;
+                    try {
+                        ppx = possibleParents(x, adjX, knowledge, y);
+                    } catch (InterruptedException e) {
+                        throw new RuntimeException(e);
+                    }
+
+                    if (adjX.size() >= d) {
+                        ChoiceGenerator generator = new ChoiceGenerator(ppx.size(), d);
+                        int[] choice;
+
+                        while ((choice = generator.next()) != null) {
+                            Set<Node> S = GraphUtils.asSet(choice, ppx);
+
+                            IndependenceResult result = null;
+                            try {
+                                result = test.checkIndependence(x, y, S);
+                            } catch (InterruptedException e) {
+                                throw new RuntimeException(e);
+                            }
+                            if (result.isIndependent() && knowledge.noEdgeRequired(x.getName(), y.getName())) {
+                                synchronized (graph_) {
+                                    graph_.removeEdge(x, y);
+                                    sepset.set(x, y, S);
+                                }
+                                break;
+                            }
+                        }
+                    }
+                }
+            });
+        } else {
+            for (int i = 0; i < nodes.size(); i++) {
+                for (int j = i + 1; j < nodes.size(); j++) {
+                    Node x = nodes.get(i);
+                    Node y = nodes.get(j);
+
+                    if (knowledge.isForbidden(x.getName(), y.getName()) && knowledge.isForbidden(y.getName(), x.getName())) {
+                        graph_.removeEdge(x, y);
+                    }
+
+                    List<Node> adjX = graph.getAdjacentNodes(x);
+                    adjX.remove(y);
+
+                    List<Node> ppx = possibleParents(x, adjX, knowledge, y);
+
+                    if (adjX.size() >= d) {
+                        ChoiceGenerator generator = new ChoiceGenerator(ppx.size(), d);
+                        int[] choice;
+
+                        while ((choice = generator.next()) != null) {
+                            Set<Node> S = GraphUtils.asSet(choice, ppx);
+
+                            IndependenceResult result = test.checkIndependence(x, y, S);
+                            if (result.isIndependent() && knowledge.noEdgeRequired(x.getName(), y.getName())) {
+                                graph_.removeEdge(x, y);
+                                sepset.set(x, y, S);
+                                break;
+                            }
+                        }
+                    }
+                }
+            }
         }
 
-        this.depth = depth;
-    }
-
-    /**
-     * Sets the knowledge for this object.
-     *
-     * @param knowledge The knowledge object to set.
-     */
-    public void setKnowledge(Knowledge knowledge) {
-        this.knowledge = new Knowledge(knowledge);
-    }
-
-    /**
-     * Returns the sepsets that were discovered in the search. A 'sepset' for test X _||_ Y | Z1,...,Zm would be
-     * {Z1,...,Zm}
-     *
-     * @return A map of sepsets discovered during the search, indexed by pairs of nodes {X, Y}.
-     */
-    public SepsetMap getSepsets() {
-        return this.sepset;
-    }
-
-    /**
-     * Sets the verbose mode.
-     *
-     * @param verbose true if verbose mode is enabled, false otherwise.
-     */
-    public void setVerbose(boolean verbose) {
-        this.verbose = verbose;
-    }
-
-    /**
-     * Returns the elapsed time of the search.
-     *
-     * @return The elapsed time of the search in milliseconds.
-     */
-    public long getElapsedTime() {
-        return elapsedTime;
-    }
-
-    /**
-     * Retrieves the list of nodes in the graph.
-     *
-     * @return A list of nodes in the graph being processed by the search algorithm.
-     */
-    @Override
-    public List<Node> getNodes() {
-        return this.test.getVariables();
-    }
-
-    /**
-     * Retrieves the list of ambiguous triples involving the given node.
-     *
-     * @param node The node for which to retrieve the ambiguous triples.
-     * @return A list of Triple objects representing the ambiguous triples involving the specified node, 
-     *         or an empty list if none are found.
-     */
-    @Override
-    public List<Triple> getAmbiguousTriples(Node node) {
-        return new ArrayList<>();
-    }
-
-    //==============================PRIVATE METHODS======================/
-
-    /**
-     * Sets the PrintStream to be used for output.
-     *
-     * @param out the PrintStream to be used for output
-     */
-    @Override
-    public void setOut(PrintStream out) {
-        this.out = out;
-    }
-
-    /**
-     * Sets the type of heuristic to be used in the PC algorithm.
-     *
-     * @param pcHeuristic The type of heuristic to be used.
-     */
-    public void setPcHeuristicType(PcCommon.PcHeuristicType pcHeuristic) {
-        this.heuristic = pcHeuristic;
-    }
-
-    /**
-     * Sets whether the stable adjacency search should be used. Default is false. Default is false. See the following
-     * reference for this:
-     * <p>
-     * Colombo, D., &amp; Maathuis, M. H. (2014). Order-independent constraint-based causal structure learning. J. Mach.
-     * Learn. Res., 15(1), 3741-3782.
-     *
-     * @param stable True iff the case.
-     */
-    public void setStable(boolean stable) {
-        this.stable = stable;
+        return freeDegree(graph_) <= d;
     }
 
     /**
      * Calculates the maximum free degree among the nodes in the given adjacency map.
      *
-     * @param adjacencies A map containing nodes as keys and sets of adjacent nodes as values.
+     * @param graph The graph.
      * @return The maximum free degree among the nodes.
      */
-    private int freeDegree(Map<Node, Set<Node>> adjacencies) {
+    private int freeDegree(Graph graph) {
         int max = 0;
 
-        for (Node x : adjacencies.keySet()) {
-            Set<Node> opposites = adjacencies.get(x);
+        for (Node n : graph.getNodes()) {
+            List<Node> opposites = graph.getAdjacentNodes(n);
 
             for (Node y : opposites) {
                 Set<Node> adjx = new LinkedHashSet<>(opposites);
@@ -624,141 +327,77 @@ public class Fas implements IFas {
     }
 
     /**
-     * Searches for adjacencies at a given depth in the graph.
+     * Sets the depth for the search process.
      *
-     * @param scores      The map of scores for each edge.
-     * @param edges       The list of edges to search over.
-     * @param test        The independence test to use.
-     * @param adjacencies The map of nodes and their adjacent nodes.
-     * @param depth       The maximum depth to search.
-     * @param mainThread  The main thread. Needed to try to interrupt the search if it takes too long.
-     * @return true if there are adjacencies at the given depth, false otherwise.
-     * @throws InterruptedException if any
+     * @param depth The maximum depth to be considered during the search.
      */
-    private boolean searchAtDepth(Map<Edge, Double> scores, List<Edge> edges, IndependenceTest test, Map<Node, Set<Node>> adjacencies, int depth, Thread mainThread) throws InterruptedException {
-        if (mainThread.isInterrupted()) {
-            return false;
-        }
-
-        // We need to write out the threading longhand here and explicitly check for InterruptedExceptions on the
-        // Futures. This is because the ForkJoinPool doesn't allow checked exceptions to be thrown from the
-        // ForkJoinTask's compute() method. This is a design flaw in the ForkJoinPool, but it is what it is.
-        // This also means that we can't catch the InterruptedException in a try/catch block for a parallel stream
-        // or a parallel stream with a collector. We have to use the ForkJoinPool and Futures. For fast independence
-        // tests, this is not a big deal, because the overhead of the ForkJoinPool is not that great. For slow
-        // independence tests, this is a big deal, because the overhead of the ForkJoinPool is significant.
-        // josephramsey 2024-12-10
-        class Task implements Callable<Boolean> {
-            private final Edge edge;
-            private final Map<Edge, Double> scores;
-            private final IndependenceTest test;
-            private final Map<Node, Set<Node>> adjacencies;
-            private final int depth;
-            private final PcCommon.PcHeuristicType heuristic;
-            private final Knowledge knowledge;
-            private final SepsetMap sepset;
-            private final Thread mainThread;
-
-            public Task(Edge edge, Map<Edge, Double> scores, IndependenceTest test, Map<Node, Set<Node>> adjacencies, int depth, PcCommon.PcHeuristicType heuristic, Knowledge knowledge, SepsetMap sepset, Thread mainThread) {
-                this.edge = edge;
-                this.scores = scores;
-                this.test = test;
-                this.adjacencies = adjacencies;
-                this.depth = depth;
-                this.heuristic = heuristic;
-                this.knowledge = knowledge;
-                this.sepset = sepset;
-                this.mainThread = mainThread;
-            }
-
-            @Override
-            public Boolean call() {
-                Node x = edge.getNode1();
-                Node y = edge.getNode2();
-
-                try {
-                    boolean b = checkSide(scores, test, adjacencies, depth, x, y, heuristic, knowledge, sepset, mainThread);
-
-                    if (!b) {
-                        checkSide(scores, test, adjacencies, depth, y, x, heuristic, knowledge, sepset, mainThread);
-                    }
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
-
-                return true;
-            }
-        }
-
-        List<Task> tasks = new ArrayList<>();
-
-        for (Edge edge : edges) {
-            tasks.add(new Task(edge, scores, test, adjacencies, depth, heuristic, knowledge, sepset, mainThread));
-        }
-
-        if (stable) {
-            int parallelism = Runtime.getRuntime().availableProcessors();
-            ForkJoinPool pool = new ForkJoinPool(parallelism);
-            List<Future<Boolean>> theseResults;
-            theseResults = pool.invokeAll(tasks);
-
-            for (Future<Boolean> future : theseResults) {
-                try {
-                    future.get();
-                } catch (InterruptedException | ExecutionException e) {
-                    TetradLogger.getInstance().log(e.getMessage());
-                    pool.shutdown();
-                    return false;
-                }
-            }
-
-            pool.shutdown();
-        } else {
-            for (Task task : tasks) {
-                task.call();
-            }
-        }
-
-        return freeDegree(adjacencies) > depth;
+    public void setDepth(int depth) {
+        this.depth = depth;
     }
 
     /**
-     * Gets the timeout value for the search operation or algorithm execution. The timeout value determines the maximum
-     * duration the process can run before it is terminated.
+     * Sets the knowledge object used in the conditional independence search process.
      *
-     * @return The timeout value for the search operation in milliseconds.
+     * @param knowledge The knowledge object that provides information about constraints or background knowledge, such
+     *                  as forbidden edges, required edges, or other conditional independencies.
      */
-    public long getTimeout() {
-        return this.timeout;
+    public void setKnowledge(Knowledge knowledge) {
+        this.knowledge = new Knowledge(knowledge);
     }
 
     /**
-     * Specifies the timeout value for the search operation or algorithm execution. The timeout value is used to
-     * determine the maximum allowed duration for the execution, preventing any operation from exceeding this limit.
+     * Retrieves the separation set map that is used during the search process to store information about conditional
+     * independence tests.
+     *
+     * @return The separation set map represented by an instance of {@code SepsetMap}, containing conditional
+     * independence information.
      */
-    public void setTimeout(long timeout) {
-        this.timeout = timeout;
+    public SepsetMap getSepsets() {
+        return this.sepset;
     }
 
     /**
-     * Represents the start time of a specific operation or process within the algorithm's execution. This variable is
-     * used to measure elapsed time or to enforce time constraints during the execution.
+     * Sets the verbosity level for the logging or debugging output of the search process.
      *
-     * @return The start time of the operation in milliseconds since epoch.
-     * 
-     * 
+     * @param verbose If true, enables verbose output; if false, disables it.
      */
-    public long getStartTime() {
-        return this.startTime;
+    public void setVerbose(boolean verbose) {
+        this.verbose = verbose;
     }
 
     /**
-     * Sets the start time for a specific operation or process.
+     * Sets the stability flag for the search process, which may determine the search strategy or algorithm's behavior.
      *
-     * @param startTime The start time in milliseconds.
+     * @param stable If true, enables a stable search strategy; otherwise, an unstable or alternative strategy is used.
      */
-    public void setStartTime(long startTime) {
-        this.startTime = startTime;
+    public void setStable(boolean stable) {
+        this.stable = stable;
+    }
+
+    /**
+     * Retrieves a list of nodes associated with the initialized independence test.
+     *
+     * @return A list of nodes derived from the variables in the associated independence test.
+     */
+    public List<Node> getNodes() {
+        return test.getVariables();
+    }
+
+    /**
+     * Retrieves the current output stream used for logging or display purposes.
+     *
+     * @return The output stream instance associated with the logging or output operations.
+     */
+    public PrintStream getOut() {
+        return out;
+    }
+
+    /**
+     * Sets the output stream to be used for logging or other display purposes.
+     *
+     * @param out The output stream to which messages or data will be printed.
+     */
+    public void setOut(PrintStream out) {
+        this.out = out;
     }
 }
-
